requirements.txt
python-telegram-bot>=21.0
openai-agents>=0.0.15
python-dotenv>=1.0.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
